+++ src/new/heuristic/main-hr_H4v1.cpp
@@ -161,7 +161,7 @@ int main(int argc, char** argv){
 	// MAIN PROCEDURE
 	DEBUG std::cerr << "Solving with breadth heuristic - vertex importance- wait!\n";
 	run_name = "Breadth_heuristic";
-	Heuristic::breadth_heuristic_2(graph);
+	Heuristic::Heuristica_4v1(graph);
 
 	// End time counting
 	std::chrono::time_point<std::chrono::steady_clock>	end = std::chrono::steady_clock::now();