+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 # changes
 
-## 0.1.0-alpha (31/01/2023)
+## 0.1.2-alpha (2023/02/08)
+FIXED a serious bug that caused core dumped (wrong index) at heuristic 4 (v1 & v2) in some graphs (G14_centrallity)
+add new graph (G13_turan_graph) at examples instances
+Replace tag name Breadth Heuristic to H4v1 for compatibility reasons
+Update my python tools at my_python_tools(these are just for my purpose), not for general public
+
+
+## 0.1.1-alpha (2023/02/05)
+Python tools added (only my purposes)
+Add new graphs (examples)
+Analysis file example (real heuristics data) added
+
+## 0.1.0-alpha (2023/01/31)
 First release! Now we have an usable code.